wip: refactoring xlsx ingest
